# apps/server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OTL Server",
    "version": "1.0.0",
    "description": "OTL PLUS main Server의 Swagger 문서입니다."
  },
  "paths": {
...
    "/api/courses": {
      "get": {
        "summary": "CourseController.getCourses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ICourse.DetailWithIsRead"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ICourse.Query"
            }
          }
        ]
      }
    },
    "/api/courses/autocomplete": {
      "get": {
        "summary": "CourseController.getCourseAutocomplete",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "string"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ICourse.AutocompleteQueryDto"
            }
          }
        ]
      }
    },
    "/api/courses/{id}": {
      "get": {
        "summary": "CourseController.getCourseById",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICourse.DetailWithIsRead"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/courses/{id}/lectures": {
      "get": {
        "summary": "CourseController.getLecturesByCourseId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ILecture.Detail"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ICourse.LectureQueryDto"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/courses/{id}/reviews": {
      "get": {
        "summary": "CourseController.getReviewByCourseId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IReview.Basic"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ICourse.ReviewQueryDto"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/courses/{id}/read": {
      "post": {
        "summary": "CourseController.readCourse",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICourse.ICourseUser.Basic"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ICourse.Basic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "old_code": {
            "type": "string"
          },
          "old_old_code": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IDepartment.Basic"
          },
          "type": {
            "type": "string"
          },
          "type_en": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_en": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "review_total_weight": {
            "type": "number"
          },
          "credit": {
            "type": "number"
          },
          "credit_au": {
            "type": "number"
          },
          "num_classes": {
            "type": "number"
          },
          "num_labs": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "old_code",
          "old_old_code",
          "department",
          "type",
          "type_en",
          "title",
          "title_en",
          "summary",
          "review_total_weight",
          "credit",
          "credit_au",
          "num_classes",
          "num_labs"
        ],
        "additionalProperties": false
      },
      "IDepartment.Basic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": ["id", "name", "name_en", "code"],
        "additionalProperties": false
      },
...
      "IProfessor.Basic": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "professor_id": {
            "type": "number"
          },
          "review_total_weight": {
            "type": "number"
          }
        },
        "required": ["name", "name_en", "professor_id", "review_total_weight"],
        "additionalProperties": false
      },
      "ICourse.DetailWithIsRead": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "old_code": {
            "type": "string"
          },
          "old_old_code": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IDepartment.Basic"
          },
          "type": {
            "type": "string"
          },
          "type_en": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "title_en": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "review_total_weight": {
            "type": "number"
          },
          "credit": {
            "type": "number"
          },
          "credit_au": {
            "type": "number"
          },
          "num_classes": {
            "type": "number"
          },
          "num_labs": {
            "type": "number"
          },
          "related_courses_prior": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICourse.Basic"
            }
          },
          "related_courses_posterior": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICourse.Basic"
            }
          },
          "professors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IProfessor.Basic"
            }
          },
          "grade": {
            "type": "number"
          },
          "load": {
            "type": "number"
          },
          "speech": {
            "type": "number"
          },
          "userspecific_is_read": {
            "type": "boolean"
          }
        },
        "required": [
          "credit",
          "credit_au",
          "department",
          "grade",
          "id",
          "load",
          "num_classes",
          "num_labs",
          "old_code",
          "old_old_code",
          "professors",
          "related_courses_posterior",
          "related_courses_prior",
          "review_total_weight",
          "speech",
          "summary",
          "title",
          "title_en",
          "type",
          "type_en",
          "userspecific_is_read"
        ],
        "additionalProperties": false
      },
      "ICourse.ICourseUser.Basic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "latest_read_datetime": {
            "type": "string",
            "format": "date-time"
          },
          "course_id": {
            "type": "number"
          },
          "user_profile_id": {
            "type": "number"
          }
        },
        "required": ["id", "latest_read_datetime", "course_id", "user_profile_id"],
        "additionalProperties": false
      },
...
      "ICourse.AutocompleteQueryDto": {
        "properties": {
          "keyword": {
            "type": "string"
          }
        },
        "type": "object",
        "required": ["keyword"]
      },
      "ICourse.Query": {
        "properties": {
          "department": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "type": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "level": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "group": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "keyword": {
            "type": "string"
          },
          "term": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "order": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "ICourse.LectureQueryDto": {
        "properties": {
          "order": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ICourse.ReviewQueryDto": {
        "properties": {
          "order": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "offset": {
            "minimum": 0,
            "type": "number"
          },
          "limit": {
            "maximum": 100,
            "type": "number",
            "minimum": 0
          }
        },
        "type": "object"
      },
...
      "IReview.Basic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "course": {
            "$ref": "#/components/schemas/ICourse.Basic"
          },
          "lecture": {
            "$ref": "#/components/schemas/ILecture.Basic"
          },
          "content": {
            "type": "string"
          },
          "like": {
            "type": "number"
          },
          "is_deleted": {
            "type": "number"
          },
          "grade": {
            "type": "number"
          },
          "load": {
            "type": "number"
          },
          "speech": {
            "type": "number"
          },
          "userspecific_is_liked": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "course",
          "lecture",
          "content",
          "like",
          "is_deleted",
          "grade",
          "load",
          "speech",
          "userspecific_is_liked"
        ],
        "additionalProperties": false
      },
      "ILecture.Basic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "title_en": {
            "type": "string"
          },
          "course": {
            "type": "number"
          },
          "old_old_code": {
            "type": "string"
          },
          "old_code": {
            "type": "string"
          },
          "class_no": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "semester": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "department": {
            "type": "number"
          },
          "department_code": {
            "type": "string"
          },
          "department_name": {
            "type": "string"
          },
          "department_name_en": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "type_en": {
            "type": "string"
          },
          "limit": {
            "type": "number"
          },
          "num_people": {
            "type": "number"
          },
          "is_english": {
            "type": "boolean"
          },
          "num_classes": {
            "type": "number"
          },
          "num_labs": {
            "type": "number"
          },
          "credit": {
            "type": "number"
          },
          "credit_au": {
            "type": "number"
          },
          "common_title": {
            "type": "string"
          },
          "common_title_en": {
            "type": "string"
          },
          "class_title": {
            "type": "string"
          },
          "class_title_en": {
            "type": "string"
          },
          "review_total_weight": {
            "type": "number"
          },
          "professors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IProfessor.Basic"
            }
          }
        },
        "required": [
          "id",
          "title",
          "title_en",
          "course",
          "old_old_code",
          "old_code",
          "class_no",
          "year",
          "semester",
          "code",
          "department",
          "department_code",
          "department_name",
          "department_name_en",
          "type",
          "type_en",
          "limit",
          "num_people",
          "is_english",
          "num_classes",
          "num_labs",
          "credit",
          "credit_au",
          "common_title",
          "common_title_en",
          "class_title",
          "class_title_en",
          "review_total_weight",
          "professors"
        ],
        "additionalProperties": false
      },
...
      "ILecture.Classtime": {
        "type": "object",
        "properties": {
          "building_code": {
            "type": "string"
          },
          "room_name": {
            "type": "string"
          },
          "classroom": {
            "type": "string"
          },
          "classroom_en": {
            "type": "string"
          },
          "classroom_short": {
            "type": "string"
          },
          "classroom_short_en": {
            "type": "string"
          },
          "day": {
            "type": "number"
          },
          "begin": {
            "type": "number"
          },
          "end": {
            "type": "number"
          }
        },
        "required": [
          "building_code",
          "room_name",
          "classroom",
          "classroom_en",
          "classroom_short",
          "classroom_short_en",
          "day",
          "begin",
          "end"
        ],
        "additionalProperties": false
      },
      "ILecture.ExamTime": {
        "type": "object",
        "properties": {
          "day": {
            "type": "number"
          },
          "str": {
            "type": "string"
          },
          "str_en": {
            "type": "string"
          },
          "begin": {
            "type": "number"
          },
          "end": {
            "type": "number"
          }
        },
        "required": ["day", "str", "str_en", "begin", "end"],
        "additionalProperties": false
      },
...
      "ILecture.Detail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "title_en": {
            "type": "string"
          },
          "course": {
            "type": "number"
          },
          "old_old_code": {
            "type": "string"
          },
          "old_code": {
            "type": "string"
          },
          "class_no": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "semester": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "department": {
            "type": "number"
          },
          "department_code": {
            "type": "string"
          },
          "department_name": {
            "type": "string"
          },
          "department_name_en": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "type_en": {
            "type": "string"
          },
          "limit": {
            "type": "number"
          },
          "num_people": {
            "type": "number"
          },
          "is_english": {
            "type": "boolean"
          },
          "num_classes": {
            "type": "number"
          },
          "num_labs": {
            "type": "number"
          },
          "credit": {
            "type": "number"
          },
          "credit_au": {
            "type": "number"
          },
          "common_title": {
            "type": "string"
          },
          "common_title_en": {
            "type": "string"
          },
          "class_title": {
            "type": "string"
          },
          "class_title_en": {
            "type": "string"
          },
          "review_total_weight": {
            "type": "number"
          },
          "professors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IProfessor.Basic"
            }
          },
          "grade": {
            "type": "number"
          },
          "load": {
            "type": "number"
          },
          "speech": {
            "type": "number"
          },
          "classtimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ILecture.Classtime"
            }
          },
          "examtimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ILecture.ExamTime"
            }
          }
        },
        "required": [
          "class_no",
          "class_title",
          "class_title_en",
          "classtimes",
          "code",
          "common_title",
          "common_title_en",
          "course",
          "credit",
          "credit_au",
          "department",
          "department_code",
          "department_name",
          "department_name_en",
          "examtimes",
          "grade",
          "id",
          "is_english",
          "limit",
          "load",
          "num_classes",
          "num_labs",
          "num_people",
          "old_code",
          "old_old_code",
          "professors",
          "review_total_weight",
          "semester",
          "speech",
          "title",
          "title_en",
          "type",
          "type_en",
          "year"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "jwtCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "sidCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "AUTH-SID"
      },
      "jwtHeaderAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "name": "accessToken"
      },
      "jwtHeaderRefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "name": "X-REFRESH-TOKEN"
      },
      "sidHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AUTH-SID"
      }
    }
  },
  "security": [
    {
      "jwtCookie": [],
      "sidCookie": [],
      "jwtHeaderAccessToken": [],
      "jwtHeaderRefreshToken": [],
      "sidHeader": []
    }
  ]
}